refactor: Refactor file paths for NavigationBar component to match new folder structure
